[ui] Add IdentitiesList component
Extracts the identities list from ExpandedIndividual
into a new component to make it reusable.
